pyproject.toml
[project]
name = "summarizers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.4",
    "dotenv>=0.9.9",
    "elevenlabs>=2.22.0",
    "faiss-cpu>=1.12.0",
    "langchain>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.0.1",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.2",
    "langchain-text-splitters>=1.0.0",
    "lxml-html-clean>=0.4.3",
    "newspaper3k>=0.2.8",
    "openai-whisper",
    "selenium>=4.38.0",
    "sentence-transformers>=5.1.2",
    "sounddevice>=0.5.3",
    "soundfile>=0.13.1",
    "streamlit>=1.51.0",
    "transformers>=4.57.1",
    "unstructured>=0.18.15",
    "yt-dlp>=2025.10.22",
]

[tool.uv.sources]
openai-whisper = { git = "https://github.com/openai/whisper.git" }

[dependency-groups]
dev = [
    "ruff>=0.14.4",
]
